agent/requirements.txt
# ==========================================
# OpenHands SDK 核心依赖
# ==========================================

# OpenHands SDK - 核心抽象层 (Agent, Conversation, LLM, Event, Tool)
openhands-sdk>=0.1.0

# OpenHands Tools - 内置工具集合 (Bash, FileEditor, TaskTracker, Terminal)
openhands-tools>=0.1.0

# OpenHands Workspace - 工作空间后端 (可选，用于本地/远程执行)
openhands-workspace>=0.1.0

# MCP (Model Context Protocol) 支持
fastmcp>=0.1.0
mcp>=1.0.0

# ==========================================
# 原有依赖
# ==========================================

# 数据验证
pydantic>=2.0.0

# YAML 解析 (用于 Skills/SOPs)
pyyaml>=6.0

# HTTP 客户端
httpx>=0.25.0

# ==========================================
# 可选依赖
# ==========================================

# 向量数据库支持 (用于 Memory 语义搜索)
# chromadb>=0.4.0

# Redis 支持 (用于分布式 Memory)
# redis>=4.5.0
